Revert "Merge DB Abstraction Layer to make OSB Database Agnostic (#1023)"
This reverts commit ed7527b9b1df95a9695fc039f028476347fdfbb0.
