+++ benchmarks/worker_coordinator/runner_test.py
@@ -24,7 +24,7 @@
 
 import pytest
 
-from osbenchmark.worker_coordinator import runners as runner
+from osbenchmark.worker_coordinator import runner
 
 bulk_index = runner.BulkIndex()
