--- osbenchmark/database/clients/milvus/__init__.py
@@ -1,7 +1,0 @@
-# SPDX-License-Identifier: Apache-2.0
-#
-# The OpenSearch Contributors require contributions made to
-# this file be licensed under the Apache-2.0 license or a
-# compatible open source license.
-
-"""Milvus client implementation."""
