+++ tests/client_test.py
@@ -34,8 +34,7 @@ import opensearchpy
 import pytest
 import urllib3.exceptions
 
-from osbenchmark.database.clients.opensearch import opensearch as client
-from osbenchmark import exceptions, doc_link
+from osbenchmark import client, exceptions, doc_link
 from osbenchmark.utils import console
 from tests import run_async
 
@@ -69,7 +68,7 @@ class OsClientFactoryTests(TestCase):
         # make a copy, so we can verify later that the factory did not modify it
         original_client_options = deepcopy(client_options)
 
-        logger = logging.getLogger("osbenchmark.database.clients.opensearch.opensearch")
+        logger = logging.getLogger("osbenchmark.client")
         with mock.patch.object(logger, "info") as mocked_info_logger:
             f = client.OsClientFactory(hosts, client_options)
         mocked_info_logger.assert_has_calls([
@@ -107,7 +106,7 @@ class OsClientFactoryTests(TestCase):
         # make a copy, so we can verify later that the factory did not modify it
         original_client_options = deepcopy(client_options)
 
-        logger = logging.getLogger("osbenchmark.database.clients.opensearch.opensearch")
+        logger = logging.getLogger("osbenchmark.client")
         with mock.patch.object(logger, "info") as mocked_info_logger:
             f = client.OsClientFactory(hosts, client_options)
         mocked_info_logger.assert_has_calls([
@@ -147,7 +146,7 @@ class OsClientFactoryTests(TestCase):
         # make a copy so we can verify later that the factory did not modify it
         original_client_options = deepcopy(client_options)
 
-        logger = logging.getLogger("osbenchmark.database.clients.opensearch.opensearch")
+        logger = logging.getLogger("osbenchmark.client")
         with mock.patch.object(logger, "info") as mocked_info_logger:
             f = client.OsClientFactory(hosts, client_options)
         mocked_info_logger.assert_has_calls([
@@ -222,7 +221,7 @@ class OsClientFactoryTests(TestCase):
         # make a copy so we can verify later that the factory did not modify it
         original_client_options = dict(client_options)
 
-        logger = logging.getLogger("osbenchmark.database.clients.opensearch.opensearch")
+        logger = logging.getLogger("osbenchmark.client")
         with mock.patch.object(logger, "info") as mocked_info_logger:
             f = client.OsClientFactory(hosts, client_options)
         mocked_info_logger.assert_has_calls([
@@ -269,7 +268,7 @@ class OsClientFactoryTests(TestCase):
             role_based_client_options
         ]
 
-        logger = logging.getLogger("osbenchmark.database.clients.opensearch.opensearch")
+        logger = logging.getLogger("osbenchmark.client")
 
         for client_options in client_options_list:
             # make a copy so we can verify later that the factory did not modify it
@@ -316,7 +315,7 @@ class OsClientFactoryTests(TestCase):
         # make a copy so we can verify later that the factory did not modify it
         original_client_options = deepcopy(client_options)
 
-        logger = logging.getLogger("osbenchmark.database.clients.opensearch.opensearch")
+        logger = logging.getLogger("osbenchmark.client")
         with mock.patch.object(logger, "info") as mocked_info_logger:
             f = client.OsClientFactory(hosts, client_options)
         mocked_info_logger.assert_has_calls([
@@ -401,7 +400,7 @@ class OsClientFactoryTests(TestCase):
         mock_session.return_value = mock_session_instance
 
         original_client_options = dict(client_options)
-        logger = logging.getLogger("osbenchmark.database.clients.opensearch.opensearch")
+        logger = logging.getLogger("osbenchmark.client")
         with mock.patch.object(logger, "info") as mocked_info_logger:
             f = client.OsClientFactory(hosts, client_options)
         mocked_info_logger.assert_has_calls([
