+++ tests/worker_coordinator/runner_test.py
@@ -31,9 +31,8 @@ from unittest import TestCase
 
 import opensearchpy
 import pytest
-from osbenchmark.database.clients.opensearch import opensearch as client
-from osbenchmark import exceptions
-from osbenchmark.worker_coordinator import runners as runner
+from osbenchmark import client, exceptions
+from osbenchmark.worker_coordinator import runner
 from tests import run_async, as_future
